refactor(hooks): 팔로잉 토글 기능 공통 커스텀 훅으로 리팩토링
